Migrate validation to Jekyll hooks (#2709)
Compared to the use of the validate.py scripts, this has the advantage of ensuring the validation is always performed.

It also gives access to enriched data, which is useful for validated generated content, such as release links. It also check links in the product description.
Note that some links had to be excluded from this check because, despite being valid, they are systematically returning an error.
A dependency to open-uri (https://github.com/ruby/open-uri) has been added to simplify the validation of HTTP links.

For this to work the priority of the ProductDataEnricher plugin has been set to normal, to allow hooks to be run before and after enrichment.

Some alternate URLs has been changed or removed due to an enforcement of the rules regarding permalinks and alternate_urls (no more dots).